fixed current tests, simplified runtime
